DEV: move select-kit plugin-api file to lib directory
