+++ app/assets/javascripts/select-kit/addon/lib/plugin-api.js
